ENH: avoid high committed memory for zonal stats calculation (#197)
The memory commited in windows was problematic, a bit worse for for
exactextract, but also for the other zonal stats calculations: ~20 GB
commited memory (not in use) per parallel process.
- on windows the processing crashed with nb_parallel=8 on an 128 GB RAM
workstation.
- not tested on linux, but linux allows overcommitting, but based on a
google warnings are emitted when a high overcommit rate is detected...

Apparently this is (mainly) caused by numpy/openblas creating a thread
pool for the 24 logical processors in each of the 8 (or 9) processes.
Setting `os.environ["OMP_NUM_THREADS"] = "1"` in the workers seems to
solve the issue.
- Reference:
https://stackoverflow.com/questions/77764228/pandas-scipy-high-commit-memory-usage-windows